Optimizing performance and adding new features
In this commit there are several improvements to the Web & Websocket Server so that they are now more performant.

In addition, a few new features have been added, including a process priority for web-based routes.

Additionally, it is no longer necessary to set a password for the SSL certificate, as this is now generated automatically at startup.