[PRVIEW] An dedicated requirement system
Adding a dedicated requirement system (for routes and socket endpoints). This also enables the precompilation of requirements for websockets.

Now custom requirements can be added and used accordingly.

Please note: As this is a preview, it may change in the future or be removed completely! Also the comments on the system classes are not final!